start of the stats file output
